// src/components/UiButton/enums.ts
export enum ButtonSize {
  S = 's',
  M = 'm',
  L = 'l',
}

export enum ButtonTheme {
  Purple = 'purple',
  PurpleBright = 'purpleBright',
  Salad = 'salad',
  White = 'white',
}

export enum ButtonState {
  Regular = 'regular',
  Hover = 'hover',
  Active = 'active',
  TextColor = 'textColor',
}